Working on a better node free space manager
Keep the free space list sorted by offset. This lets us easily merge free blocks when they are put in the list. This should reduce the need for node defragmentation when records have varying sizes.

+++ src/header.h
@@ -77,7 +77,7 @@ struct NodeHeader {
     std::uint16_t cell_start {};
     std::uint16_t free_start {};
     std::uint16_t free_total {};
-    std::uint8_t frag_count {};
+    std::uint16_t frag_count {};
     bool is_external {true};
 };
